BaseTools/UPT: Porting UPT Tool from Python2 to Python3
Cc: Liming Gao <liming.gao@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1

+++ BaseTools/Source/Python/UPT/Parser/DecParserMisc.py
@@ -151,7 +151,7 @@ def IsValidNumValUint8(Token):
     else:
         Base = 10
     try:
-        TokenValue = long(Token, Base)
+        TokenValue = int(Token, Base)
     except BaseException:
         Valid, Cause = IsValidLogicalExpr(Token, True)
         if Cause:
@@ -262,7 +262,7 @@ def IsValidPcdDatum(Type, Value):
                 Value = Value.lstrip('0')
                 if not Value:
                     return True, ""
-            Value = long(Value, 0)
+            Value = int(Value, 0)
             TypeLenMap = {
                 #
                 # 0x00 - 0xff
